1.更改了部分表字段类型 2.增加jwt和md5工具类 3.编写造简单数据测试类

+++ src/main/java/com/water/quality/pojo/entity/Notice.java
@@ -2,19 +2,20 @@ package com.water.quality.pojo.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.time.LocalDateTime;
-import java.io.Serializable;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * <p>
  * 消息通知表
  * </p>
  *
- * @author 
+ * @author
  * @since 2022-02-10
  */
 @Data
@@ -25,13 +26,13 @@ public class Notice implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "0新节点通知 1污染记录通知")
     private Integer noticeType;
 
     @ApiModelProperty(value = "监测点id或污染记录id")
-    private Integer siteId;
+    private Long siteId;
 
     @ApiModelProperty(value = "通知标题")
     private String title;
@@ -46,10 +47,10 @@ public class Notice implements Serializable {
     private LocalDateTime updateTime;
 
     @ApiModelProperty(value = "创建人id")
-    private Integer creatorId;
+    private Long creatorId;
 
     @ApiModelProperty(value = "更新人id")
-    private Integer editorId;
+    private Long editorId;
 
 
 }
